requirements.txt
breadability==0.1.20
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
decorator==4.4.2
docopt==0.6.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
lxml==4.5.0
MarkupSafe==1.1.1
networkx==2.4
nltk==3.5
num2words==0.5.10
numpy==1.18.4
pandas==1.0.3
pdfminer.six==20200402
pycountry==19.8.18
pycryptodome==3.9.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.5.7
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sortedcontainers==2.1.0
sumy==0.8.1
tqdm==4.46.0
urllib3==1.25.9
Werkzeug==1.0.1
